Add everforest colorscheme
Could have picked any number of schemes here. I went with a theme that I
know is popular in the community but that is also one that I wouldn't
normally pick myself. Good to separate myself from this project a little
bit to that it doesn't become a reimagining of my config too much